@@ -5,21 +5,24 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
-import kotlinx.android.synthetic.main.activity_main.*
+import com.cbellmont.ejemploandroid8.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
 
     companion object {
-        const val TAG = "TAG"
+        const val TAG = "Hola"
     }
+
+    lateinit var binding : ActivityMainBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
-        cargarPreferencias()?.let { editText.setText(it) }
+        cargarPreferencias()?.let { binding.editText.setText(it) }
 
-        editText.addTextChangedListener(object : TextWatcher {
+        binding.editText.addTextChangedListener(object : TextWatcher {
             override fun afterTextChanged(s: Editable?) {
                 guardarPreferencias(s.toString())
             }
@@ -41,6 +44,11 @@ class MainActivity : AppCompatActivity() {
 
     private fun guardarPreferencias(string : String) {
         val sharedPref = getPreferences(Context.MODE_PRIVATE)
+        /* Lo de abajo es una manera más elegante de poner esto mismo.
+        var sharedPrefEditable = sharedPref.edit()
+        sharedPrefEditable.putString("NOMBRE", "asdfasdf")
+        sharedPrefEditable.commit() */
+
         with (sharedPref.edit()) {
             putString(TAG, string)
             commit()
